+++ Cocoa/Plug-Ins/Template Editor/ElementUBYT.m
@@ -46,7 +46,7 @@
 	char cstr[256];
 	char *endPtr = cstr + 255;
 	strncpy(cstr, [str cStringUsingEncoding:NSMacOSRomanStringEncoding], 255);
-	value = strtoul(cstr, &endPtr, 10);
+	value = (UInt8)strtoul(cstr, &endPtr, 10);
 }
 
 @end
